Removed confighelper to allow config to be loaded alone
